# Qualifiers #1751: refactoring suggestion