suggestion for #1751 

IMO separating these concerns feels less tangly, easier on the mind, even if it is a bit more code. I think it's also quite well optimized for performance as originally intended.