@stevenschlansker 2020-09-14
Did you have a chance to compare performance before / after / vs my change?