@leaumar 2020-09-14
No, sorry. I can try to do that tomorrow if you don't do it. My focus was on code simplicity, I just tried to avoid doing anything that would be an obvious step backward in performance.