@leaumar 2020-09-14
Before refactor:
```
Benchmark                                     Mode  Cnt      Score      Error  Units
QualifiersBenchmark.getQualifiersAnnotated    avgt    5    225.286 ±   17.535  ns/op
QualifiersBenchmark.getQualifiersUnannotated  avgt    5    135.407 ±    6.985  ns/op
QualifiersBenchmark.mapQualifiedBean          avgt    5  32375.018 ± 1512.807  ns/op
QualifiersBenchmark.mapUnqualifiedBean        avgt    5  34898.519 ± 7136.024  ns/op
```

After refactor:
```
Benchmark                                     Mode  Cnt      Score      Error  Units
QualifiersBenchmark.getQualifiersAnnotated    avgt    5    222.257 ±   10.369  ns/op
QualifiersBenchmark.getQualifiersUnannotated  avgt    5    140.961 ±   10.855  ns/op
QualifiersBenchmark.mapQualifiedBean          avgt    5  32894.482 ± 2735.528  ns/op
QualifiersBenchmark.mapUnqualifiedBean        avgt    5  32898.203 ± 2078.309  ns/op
```

Seems to me the difference is within the margin of error, so no loss of performance?