@stevenschlansker 2020-09-15
Unfortunately I don't think the QualifiersBenchmark actually covers this refactoring - you'll notice it changes argument factory code, while the qualifier benchmark only tests the Qualifiers code itself.