@leaumar 2020-09-15
The only way it touches argument factory code is renaming the method you added in your original PR (which was one of my other review remarks), so I don't think it matters for the concern of refactoring just a part of your original PR?