@stevenschlansker 2020-09-16
I tried to make some comparisons but even running for 2h the results were inconclusive.
I guess I'll save this for last, get the rest merged in, then re-test and re-profile.